Null entity validation

+++ Implementation/ServiceImplementations/BaseDeleteService.cs
@@ -1,5 +1,6 @@
 ﻿using Data;
 using Domen;
+using FluentValidation;
 using MikroServisProizvod.Application.DefaultServices;
 using System;
 using System.Collections.Generic;
@@ -18,6 +19,13 @@ namespace MikroServisProizvod.Implementation.ServiceImplementations
 
         public void Delete(long id)
         {
+            var entity = GenericRepository.FirstOrDefault(x => x.Id == id);
+
+            if (entity is null) // prvo proverimo da li entity za brisanje postoji u bazi
+            {
+                throw new ValidationException($"Nepostojeci {typeof(TEntity).Name.ToLower()} poslat na brisanje.");
+            }
+
             GenericRepository.Delete(id);
         }
     }
